requirements.txt
# --extra-index-url https://download.pytorch.org/whl/cu118
torch
torchvision

pillow
gradio
transformers
supervision